Adding corpus tests in two places.
The original files are in corpus_with_actions/, the files currently in
corpus/ are without  @foo{} blocks because those don't quite play nicely with
the parser as it stands.

Most of the problems so far appear to be with commenting not being allowed in
places where it should be.

I'm committing this before removing all of the token TERMINAL nonsense, because
those clutter up the parse tree and it's already cluttered enough as it is.

Also having EOF as its own terminal slowed things down noticeably.