Zow. Ignoring comments and actions, ~90% of the grammars parse cleanly.
I've completely removed the comment code from the core ANTLR grammar for the
time being so that I can concentrate on fixing the broken rules, then adding
the comment stuff in where appropriate. The original grammar only allowed
comments in a few places anyway, and not where the existing grammar subsets
did in any case.